Ce n'était toujours pas parfait, alors je l'ai fait moi-même!

+++ client/public_html/js/app.js
@@ -269,14 +269,14 @@ $(document).ready(function() {
                     'data':JSON.stringify(vin),
                     'contentType':'application/json'
                 }).done(function(data) {
-                    if(data) {
-                        //Actualiser la liste de tous les vins
-                        showWines();
-                        
-                        reportError('Le vin a bien été modifié.','success');
-                    } else {
-                        reportError('Désolé, Impossible de modifier ce vin!','error');
-                    }
+                        if(data) {
+                            //Actualiser la liste de tous les vins
+                            showWines();
+
+                            reportError('Le vin a bien été modifié.','success');
+                        } else {
+                            reportError('Désolé, Impossible de modifier ce vin!','error');
+                        }
                 }).fail(function() {
                     reportError('Désolé, Impossible de modifier ce vin!','error');
                 });
@@ -293,25 +293,25 @@ $(document).ready(function() {
     $('#btDeleteWine').on('click', function() {
         //Retirer la notification d'erreur
         removeError();
-        
         let vin = getFormData();
-                
-        $.ajax({
-            'method':'DELETE',
-            'url':API_URL+'/wines/'+vin.id,
-        }).done(function(data){
-            if(data) {
-                //Actualiser la liste de tous les vins
-                showWines();
-                
-                reportError('Le vin a bien été supprimé.','success');
-            } else {
-                reportError('Désolé, Impossible de supprimer ce vin!','error');
-            }
-        }).fail(function() {
-            reportError('Désolé, Impossible de supprimer ce vin!','error');
-        });
         
+        if(confirm('Confirmer la suppression ?')){        
+            $.ajax({
+                'method':'DELETE',
+                'url':API_URL+'/wines/'+vin.id,
+            }).done(function(data) {
+                if(data) {
+                //Actualiser la liste de tous les vins
+                    showWines();
+                    reportError('Le vin a bien été supprimé.','success');
+                }else {
+
+                    reportError('Désolé, Impossible de supprimer ce vin!','error');
+                }
+            }).fail(function() {
+                        reportError('Désolé, Impossible de supprimer ce vin!','error'); 
+            });
+        }
         //Annuler l'envoi du formulaire
         event.preventDefault();
         return false;
